requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
certifi==2025.6.15
charset-normalizer==3.4.2
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
distro==1.9.0
Django==5.2.3
fonttools==4.58.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
kiwisolver==1.4.8
Markdown==3.8.2
matplotlib==3.8.4
numpy==1.26.4
openai==1.92.2
packaging==25.0
pillow==11.2.1
pydantic==2.11.7
pydantic_core==2.33.2
pyparsing==3.0.9
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-telegram-bot==22.1
requests==2.32.4
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
wheel==0.45.1
whitenoise==6.8.2
fpdf2
python-docx
pandas
xlsxwriter